moving constants to a separate folder
