+++ .env.example
@@ -1,0 +1,1 @@
+GOERLI_RPC_URL=
