--- src/constants.ts
@@ -1,11 +1,0 @@
-export const ConfigUrl = {
-  devnet: "https://chainbridge-assets-stage.s3.us-east-2.amazonaws.com/shared-config-dev.json",
-  testnet: "https://chainbridge-assets-stage.s3.us-east-2.amazonaws.com/shared-config-test.json",
-  mainnet: "https://sygma-assets-mainnet.s3.us-east-2.amazonaws.com/shared-config-mainnet.json"
-}
-
-// chain id -> provider url
-export const chainIdToRpc = { 
-  5: "https://goerli.infura.io/v3/"
-}
-
