+++ src/constants/index.ts
@@ -1,0 +1,2 @@
+export { chainIdToRpc } from "./rpc";
+export { SharedConfig } from "./sharedConfig";
