+++ src/constants/rpc.ts
@@ -1,0 +1,5 @@
+// chain id -> provider url
+export const chainIdToRpc = {
+    5: process.env.GOERLI_RPC_URL,
+}
+
