+++ src/constants/sharedConfig.ts
@@ -1,0 +1,5 @@
+export const SharedConfig = {
+    devnet: "https://chainbridge-assets-stage.s3.us-east-2.amazonaws.com/shared-config-dev.json",
+    testnet: "https://chainbridge-assets-stage.s3.us-east-2.amazonaws.com/shared-config-test.json",
+    mainnet: "https://sygma-assets-mainnet.s3.us-east-2.amazonaws.com/shared-config-mainnet.json"
+};
