+++ types/bridgePausing.d.ts
@@ -1,2 +1,2 @@
-export {};
+import 'dotenv/config';
 //# sourceMappingURL=bridgePausing.d.ts.map
