--- types/constants.d.ts
@@ -1,9 +1,0 @@
-export declare const ConfigUrl: {
-    devnet: string;
-    testnet: string;
-    mainnet: string;
-};
-export declare const chainIdToRpc: {
-    5: string;
-};
-//# sourceMappingURL=constants.d.ts.map
