+++ types/constants/rpc.d.ts
@@ -1,0 +1,4 @@
+export declare const chainIdToRpc: {
+    5: string | undefined;
+};
+//# sourceMappingURL=rpc.d.ts.map
