+++ types/constants/sharedConfig.d.ts
@@ -1,0 +1,6 @@
+export declare const SharedConfig: {
+    devnet: string;
+    testnet: string;
+    mainnet: string;
+};
+//# sourceMappingURL=sharedConfig.d.ts.map
